Extract shared sheet chrome helpers
